Fix uint8 (#21)
* Make TLC Count uint16

* fixup

* Update TLC5955.cpp

---------

+++ TLC5955.h
@@ -84,7 +84,7 @@ uint32_t getSpiBaudRate();
 void setGsclkFreq(uint32_t new_gsclk_frequency);
 uint32_t getGsclkFreq();
 
-static const uint8_t _tlc_count; // This
+static const int32_t _tlc_count; // This
 static const uint8_t COLOR_CHANNEL_COUNT = 3;
 static const uint8_t LEDS_PER_CHIP = 16;
 static bool enforce_max_current;
